Bootstrap+Packages: Simply bootstrap script and package file.
Remove the option for dialogs and make the bootstrapping process
implicity automatic - having dialogs as an install method made
the script far too long and I much prefer it this way.

I've also decided to remove the checks for nvidia and virtualbox
drivers as I believe this is something seperate, and not to be
handled by the bootstrap script itself.