packages+fish: Fix gpg-signing issue.
Before it was not possible to sign commits via magit
or the terminal. Exporting 'GPG_TTY' as an enviroment
variable allows me to commit from the terminal, as gpg
looks for this.

Whilst not idea I'll have to use the gtk2 pinentry dialog
to sign commits from within magit.

I take no credit for this:

bonfacemunyoki.com/post/2019-11-25-fixing-gpg-signing-in-magit/

Using the above article resolved my issues.

Signed-off-by: Lewis Cook <vulcan@wired.sh>