Scripts: 'gl-port' Add '-e' and '-l' flags.
Now we can specify the limit of logs to return,
and the given entry for cases when we want to merge
multiple commits in one.

For instance, we update 'foo' in svn and have yet to
push the updates to the git repo. During this time the
'PORTREVISION' is bumped so now have two commits to push.
Oppose to two individual commits we can merge them with the
following:

gl-port -u -p foo -c sysutils -l 3 -e 2

Signed-off-by: Lewis Cook <vulcan@wired.sh>